1. 주간 Top 게시물 조회시 hotscore 함께 반환

+++ src/main/java/net/whgkswo/excuse_bundle/entities/posts/hotscore/PostWithHotScoreDto.java
@@ -2,5 +2,5 @@ package net.whgkswo.excuse_bundle.entities.posts.hotscore;
 
 import net.whgkswo.excuse_bundle.entities.posts.core.entity.Post;
 
-public record PostWithHotScoreDto(Post post, double hotScore) {
+public record PostWithHotScoreDto(Post post, int hotScore) {
 }
